BRENDASearch

search for BRENDA tissue name updated 02/2024

BRENDASearch(x, file = BRENDA, contain_all = TRUE)

Arguments

x

A string or character vector to search for. Multiple search terms are separated by space.

file

The BRENDA data frame, has columns ID and Name.

contain_all

if TRUE, will only return the results contain all search terms. if FALSE, will return results contain any of the given pattern.

Value

matrix including search result

Examples

# search for results that contain all of "HEK" "293" and "T":
BRENDASearch("HEK 293 T", contain_all = TRUE)
#>               ID                      Name
#> 2180 BTO:0002181             HEK-293T cell
#> 3346 BTO:0003347    HEK-293 Tet-On 3G cell
#> 5023 BTO:0005025         HEK-293-TrkB cell
#> 5027 BTO:0005029            HEK-293FT cell
#> 5546 BTO:0005548            HEK-293ET cell
#> 6320 BTO:0006322 Jump-In-T-REx HEK293 cell
#> 6326 BTO:0006328          HEK-293T/17 cell
#> 6511 BTO:0006513          HEK-293-AT1 cell
BRENDASearch(c("HEK", "293", "T"), contain_all = TRUE)
#>               ID                      Name
#> 2180 BTO:0002181             HEK-293T cell
#> 3346 BTO:0003347    HEK-293 Tet-On 3G cell
#> 5023 BTO:0005025         HEK-293-TrkB cell
#> 5027 BTO:0005029            HEK-293FT cell
#> 5546 BTO:0005548            HEK-293ET cell
#> 6320 BTO:0006322 Jump-In-T-REx HEK293 cell
#> 6326 BTO:0006328          HEK-293T/17 cell
#> 6511 BTO:0006513          HEK-293-AT1 cell
BRENDASearch(c("HEK 293", "T"), contain_all = TRUE)
#>               ID                      Name
#> 2180 BTO:0002181             HEK-293T cell
#> 3346 BTO:0003347    HEK-293 Tet-On 3G cell
#> 5023 BTO:0005025         HEK-293-TrkB cell
#> 5027 BTO:0005029            HEK-293FT cell
#> 5546 BTO:0005548            HEK-293ET cell
#> 6320 BTO:0006322 Jump-In-T-REx HEK293 cell
#> 6326 BTO:0006328          HEK-293T/17 cell
#> 6511 BTO:0006513          HEK-293-AT1 cell

# search for results that contain any of "HEK", "SUM" or "HeLa":
BRENDASearch("HEK SUM HeLa", contain_all = FALSE)
#>               ID                      Name
#> 567  BTO:0000567                 HeLa cell
#> 568  BTO:0000568              HeLa-S3 cell
#> 3179 BTO:0003180              HeLa-80 cell
#> 3586 BTO:0003587       HeLa-MAGI-CCR5 cell
#> 3587 BTO:0003588            HeLa-MAGI cell
#> 3601 BTO:0003602 HeLa GFP-histone H2B cell
#> 5028 BTO:0005030             HeLa-229 cell
#> 5036 BTO:0005038    HeLa-LTRHIV-1-Luc cell
#> 6089 BTO:0006091           HeLa-HL3T1 cell
#> 4926 BTO:0004928            SUM-159PT cell
#> 5385 BTO:0005387            SUM-102PT cell
#> 5386 BTO:0005388              SUM-149 cell
#> 5387 BTO:0005389             SUM-44PE cell
#> 5388 BTO:0005390             SUM-52PE cell
#> 5389 BTO:0005391              SUM-225 cell
#> 5390 BTO:0005392              SUM-229 cell
#> 5391 BTO:0005393              SUM-190 cell
#> 5392 BTO:0005394              SUM-185 cell
#> 5473 BTO:0005475            SUM-149PT cell
#> 5474 BTO:0005476            SUM-185PE cell
#> 5475 BTO:0005477            SUM-190PT cell
#> 5476 BTO:0005478           SUM-225CWN cell
#> 5477 BTO:0005479            SUM-229PE cell
#> 5930 BTO:0005932                  SUM cell
#> 5931 BTO:0005933          SUM-1315MO2 cell
#> 8    BTO:0000007              HEK-293 cell
#> 2180 BTO:0002181             HEK-293T cell
#> 2523 BTO:0002524             HEK-293A cell
#> 2731 BTO:0002732          HEK-293CymR cell
#> 2973 BTO:0002974         HEK-293-EBNA cell
#> 3031 BTO:0003032             HEK-293F cell
#> 3032 BTO:0003033             HEK-293H cell
#> 3181 BTO:0003182                 NHEK cell
#> 3262 BTO:0003263            HEK-AD293 cell
#> 3346 BTO:0003347    HEK-293 Tet-On 3G cell
#> 3467 BTO:0003468            HEK-293B2 cell
#> 3614 BTO:0003615    HEK-293 PEAKrapid cell
#> 4073 BTO:0004075           D1-HEK-293 cell
#> 4790 BTO:0004792               RHEK-1 cell
#> 5023 BTO:0005025         HEK-293-TrkB cell
#> 5027 BTO:0005029            HEK-293FT cell
#> 5546 BTO:0005548            HEK-293ET cell
#> 5837 BTO:0005839           HEK-293VnR cell
#> 6095 BTO:0006097                 HEKa cell
#> 6320 BTO:0006322 Jump-In-T-REx HEK293 cell
#> 6326 BTO:0006328          HEK-293T/17 cell
#> 6420 BTO:0006422       HEK-293-APPswe cell
#> 6451 BTO:0006453             HEK-293S cell
#> 6511 BTO:0006513          HEK-293-AT1 cell
BRENDASearch(c("HEK", "SUM", "HeLa"), contain_all = FALSE)
#>               ID                      Name
#> 567  BTO:0000567                 HeLa cell
#> 568  BTO:0000568              HeLa-S3 cell
#> 3179 BTO:0003180              HeLa-80 cell
#> 3586 BTO:0003587       HeLa-MAGI-CCR5 cell
#> 3587 BTO:0003588            HeLa-MAGI cell
#> 3601 BTO:0003602 HeLa GFP-histone H2B cell
#> 5028 BTO:0005030             HeLa-229 cell
#> 5036 BTO:0005038    HeLa-LTRHIV-1-Luc cell
#> 6089 BTO:0006091           HeLa-HL3T1 cell
#> 4926 BTO:0004928            SUM-159PT cell
#> 5385 BTO:0005387            SUM-102PT cell
#> 5386 BTO:0005388              SUM-149 cell
#> 5387 BTO:0005389             SUM-44PE cell
#> 5388 BTO:0005390             SUM-52PE cell
#> 5389 BTO:0005391              SUM-225 cell
#> 5390 BTO:0005392              SUM-229 cell
#> 5391 BTO:0005393              SUM-190 cell
#> 5392 BTO:0005394              SUM-185 cell
#> 5473 BTO:0005475            SUM-149PT cell
#> 5474 BTO:0005476            SUM-185PE cell
#> 5475 BTO:0005477            SUM-190PT cell
#> 5476 BTO:0005478           SUM-225CWN cell
#> 5477 BTO:0005479            SUM-229PE cell
#> 5930 BTO:0005932                  SUM cell
#> 5931 BTO:0005933          SUM-1315MO2 cell
#> 8    BTO:0000007              HEK-293 cell
#> 2180 BTO:0002181             HEK-293T cell
#> 2523 BTO:0002524             HEK-293A cell
#> 2731 BTO:0002732          HEK-293CymR cell
#> 2973 BTO:0002974         HEK-293-EBNA cell
#> 3031 BTO:0003032             HEK-293F cell
#> 3032 BTO:0003033             HEK-293H cell
#> 3181 BTO:0003182                 NHEK cell
#> 3262 BTO:0003263            HEK-AD293 cell
#> 3346 BTO:0003347    HEK-293 Tet-On 3G cell
#> 3467 BTO:0003468            HEK-293B2 cell
#> 3614 BTO:0003615    HEK-293 PEAKrapid cell
#> 4073 BTO:0004075           D1-HEK-293 cell
#> 4790 BTO:0004792               RHEK-1 cell
#> 5023 BTO:0005025         HEK-293-TrkB cell
#> 5027 BTO:0005029            HEK-293FT cell
#> 5546 BTO:0005548            HEK-293ET cell
#> 5837 BTO:0005839           HEK-293VnR cell
#> 6095 BTO:0006097                 HEKa cell
#> 6320 BTO:0006322 Jump-In-T-REx HEK293 cell
#> 6326 BTO:0006328          HEK-293T/17 cell
#> 6420 BTO:0006422       HEK-293-APPswe cell
#> 6451 BTO:0006453             HEK-293S cell
#> 6511 BTO:0006513          HEK-293-AT1 cell
BRENDASearch(c("HEK SUM", "HeLa"), contain_all = FALSE)
#>               ID                      Name
#> 567  BTO:0000567                 HeLa cell
#> 568  BTO:0000568              HeLa-S3 cell
#> 3179 BTO:0003180              HeLa-80 cell
#> 3586 BTO:0003587       HeLa-MAGI-CCR5 cell
#> 3587 BTO:0003588            HeLa-MAGI cell
#> 3601 BTO:0003602 HeLa GFP-histone H2B cell
#> 5028 BTO:0005030             HeLa-229 cell
#> 5036 BTO:0005038    HeLa-LTRHIV-1-Luc cell
#> 6089 BTO:0006091           HeLa-HL3T1 cell
#> 4926 BTO:0004928            SUM-159PT cell
#> 5385 BTO:0005387            SUM-102PT cell
#> 5386 BTO:0005388              SUM-149 cell
#> 5387 BTO:0005389             SUM-44PE cell
#> 5388 BTO:0005390             SUM-52PE cell
#> 5389 BTO:0005391              SUM-225 cell
#> 5390 BTO:0005392              SUM-229 cell
#> 5391 BTO:0005393              SUM-190 cell
#> 5392 BTO:0005394              SUM-185 cell
#> 5473 BTO:0005475            SUM-149PT cell
#> 5474 BTO:0005476            SUM-185PE cell
#> 5475 BTO:0005477            SUM-190PT cell
#> 5476 BTO:0005478           SUM-225CWN cell
#> 5477 BTO:0005479            SUM-229PE cell
#> 5930 BTO:0005932                  SUM cell
#> 5931 BTO:0005933          SUM-1315MO2 cell
#> 8    BTO:0000007              HEK-293 cell
#> 2180 BTO:0002181             HEK-293T cell
#> 2523 BTO:0002524             HEK-293A cell
#> 2731 BTO:0002732          HEK-293CymR cell
#> 2973 BTO:0002974         HEK-293-EBNA cell
#> 3031 BTO:0003032             HEK-293F cell
#> 3032 BTO:0003033             HEK-293H cell
#> 3181 BTO:0003182                 NHEK cell
#> 3262 BTO:0003263            HEK-AD293 cell
#> 3346 BTO:0003347    HEK-293 Tet-On 3G cell
#> 3467 BTO:0003468            HEK-293B2 cell
#> 3614 BTO:0003615    HEK-293 PEAKrapid cell
#> 4073 BTO:0004075           D1-HEK-293 cell
#> 4790 BTO:0004792               RHEK-1 cell
#> 5023 BTO:0005025         HEK-293-TrkB cell
#> 5027 BTO:0005029            HEK-293FT cell
#> 5546 BTO:0005548            HEK-293ET cell
#> 5837 BTO:0005839           HEK-293VnR cell
#> 6095 BTO:0006097                 HEKa cell
#> 6320 BTO:0006322 Jump-In-T-REx HEK293 cell
#> 6326 BTO:0006328          HEK-293T/17 cell
#> 6420 BTO:0006422       HEK-293-APPswe cell
#> 6451 BTO:0006453             HEK-293S cell
#> 6511 BTO:0006513          HEK-293-AT1 cell