template for creating a metadata list for an OmicSignature R6 object updated 09/2020

template for creating a metadata list for an OmicSignature R6 object updated 09/2020

createMetadata(
  signature_name,
  organism,
  phenotype = "unknown",
  covariates = "none",
  platform,
  direction_type,
  sample_type = NULL,
  signature_collection = NULL,
  author = NULL,
  year = NULL,
  PMID = NULL,
  keywords = NULL,
  description = NULL,
  category_num = NULL,
  logfc_cutoff = NULL,
  p_value_cutoff = NULL,
  adj_p_cutoff = NULL,
  score_cutoff = NULL,
  cutoff_description = NULL,
  ...
)

Arguments

signature_name
name of the signature.

organism
organism. e.g. "Homo Sapiens", "Mus Musculus".

phenotype
the phenotype of the signature. e.g. "Gene KO", "Ketodiet", "Parkinson disease", "unknown".

covariates
optional but highly recommended. any covariates used in the study. e.g. a signature of aging with covariate "gender".

platform
GEO platform. e.g. "GPL11154" is for Illumina HiSeq 2000 Homo sapiens, "GPL570" is for Affymetrix Human Genome U133 Plus 2.0 Array

direction_type
the direction information of the signature. "uni" or "uni-directional" if signature has only one direction or no direction infomation. "bi" or "bi-directional" if signature contains "up" and "down" regulated features. "multi" or "multi-directional" if the signature contains more categories.

sample_type
optional but highly recommended. a cell line or tissue from BRENDA ontology.

signature_collection
optional. collection of which the signature belongs to.

author
optional. the author name if the signature is from a published article.

year
optional. the year when the signature was created or published.

PMID
optional. the PubMed ID if the signature is from a published article.

keywords
optional. key words for the signature. examples are "longevity", "perturbation". "drug".

description
optional. free text to describe the signature.

category_num
optional, specifically used for multi-directional signature to specify how many categories or class the signature contains.

logfc_cutoff
optional. log fold change cutoff used to generate the signature, if applicable.

p_value_cutoff
optional. p value cutoff used to generate the signature, if applicable.

adj_p_cutoff
optional. adjusted p-value, e.g. fdr, cutoff used to generate the signature, if applicable.

score_cutoff
optional. score cutoff used to generate the signature, if applicable.

cutoff_description
optional. discription of the cutoff, if applicable.

...
additional user-defined metadata fields.

Value

a metadata list to create an OmicSignature R6 object