searchSampleType

search for BRENDA tissue name updated 02/2024

searchSampleType(x, file = BRENDA, contain_all = TRUE)

Arguments

x

A string or character vector to search for. Multiple search terms are separated by space.

file

The BRENDA data frame, has columns ID and Name.

contain_all

if TRUE, will only return the results contain all search terms. if FALSE, will return results contain any of the given pattern.

Value

matrix including search result

Examples

# search for results that contain all of "HEK" "293" and "T":
searchSampleType("HEK 293 T", contain_all = TRUE)
#>               ID                      Name
#> 2180 BTO:0002181             HEK-293T cell
#> 3346 BTO:0003347    HEK-293 Tet-On 3G cell
#> 5021 BTO:0005025         HEK-293-TrkB cell
#> 5025 BTO:0005029            HEK-293FT cell
#> 5544 BTO:0005548            HEK-293ET cell
#> 6318 BTO:0006322 Jump-In-T-REx HEK293 cell
#> 6324 BTO:0006328          HEK-293T/17 cell
#> 6509 BTO:0006513          HEK-293-AT1 cell
searchSampleType(c("HEK", "293", "T"), contain_all = TRUE)
#>               ID                      Name
#> 2180 BTO:0002181             HEK-293T cell
#> 3346 BTO:0003347    HEK-293 Tet-On 3G cell
#> 5021 BTO:0005025         HEK-293-TrkB cell
#> 5025 BTO:0005029            HEK-293FT cell
#> 5544 BTO:0005548            HEK-293ET cell
#> 6318 BTO:0006322 Jump-In-T-REx HEK293 cell
#> 6324 BTO:0006328          HEK-293T/17 cell
#> 6509 BTO:0006513          HEK-293-AT1 cell
searchSampleType(c("HEK 293", "T"), contain_all = TRUE)
#>               ID                      Name
#> 2180 BTO:0002181             HEK-293T cell
#> 3346 BTO:0003347    HEK-293 Tet-On 3G cell
#> 5021 BTO:0005025         HEK-293-TrkB cell
#> 5025 BTO:0005029            HEK-293FT cell
#> 5544 BTO:0005548            HEK-293ET cell
#> 6318 BTO:0006322 Jump-In-T-REx HEK293 cell
#> 6324 BTO:0006328          HEK-293T/17 cell
#> 6509 BTO:0006513          HEK-293-AT1 cell

# search for results that contain any of "HEK", "SUM" or "HeLa":
searchSampleType("HEK SUM HeLa", contain_all = FALSE)
#>               ID                                  Name
#> 567  BTO:0000567                             HeLa cell
#> 568  BTO:0000568                          HeLa-S3 cell
#> 2771 BTO:0002772                                 chela
#> 3179 BTO:0003180                          HeLa-80 cell
#> 3586 BTO:0003587                   HeLa-MAGI-CCR5 cell
#> 3587 BTO:0003588                        HeLa-MAGI cell
#> 3601 BTO:0003602             HeLa GFP-histone H2B cell
#> 5026 BTO:0005030                         HeLa-229 cell
#> 5034 BTO:0005038                HeLa-LTRHIV-1-Luc cell
#> 5995 BTO:0005999                           chelate leg
#> 6087 BTO:0006091                       HeLa-HL3T1 cell
#> 25   BTO:0000024                              abomasum
#> 273  BTO:0000272                     colon transversum
#> 349  BTO:0000348                                omasum
#> 362  BTO:0000361                    stratum granulosum
#> 436  BTO:0000435                      stratum spinosum
#> 615  BTO:0000615                       corpus callosum
#> 1712 BTO:0001713                                dorsum
#> 2017 BTO:0002018          corpus cavernosum clitoridis
#> 2018 BTO:0002019               corpus cavernosum penis
#> 3676 BTO:0003677                     chorion frondosum
#> 4133 BTO:0004136                         Kasumi-1 cell
#> 4422 BTO:0004425          stratum granulosum cerebelli
#> 4924 BTO:0004928                        SUM-159PT cell
#> 5383 BTO:0005387                        SUM-102PT cell
#> 5384 BTO:0005388                          SUM-149 cell
#> 5385 BTO:0005389                         SUM-44PE cell
#> 5386 BTO:0005390                         SUM-52PE cell
#> 5387 BTO:0005391                          SUM-225 cell
#> 5388 BTO:0005392                          SUM-229 cell
#> 5389 BTO:0005393                          SUM-190 cell
#> 5390 BTO:0005394                          SUM-185 cell
#> 5450 BTO:0005454                     corpus cavernosum
#> 5471 BTO:0005475                        SUM-149PT cell
#> 5472 BTO:0005476                        SUM-185PE cell
#> 5473 BTO:0005477                        SUM-190PT cell
#> 5474 BTO:0005478                       SUM-225CWN cell
#> 5475 BTO:0005479                        SUM-229PE cell
#> 5499 BTO:0005503       corpus cavernosum smooth muscle
#> 5928 BTO:0005932                              SUM cell
#> 5929 BTO:0005933                      SUM-1315MO2 cell
#> 6264 BTO:0006268 organum vasculosum laminae terminalis
#> 8    BTO:0000007                          HEK-293 cell
#> 2180 BTO:0002181                         HEK-293T cell
#> 2523 BTO:0002524                         HEK-293A cell
#> 2731 BTO:0002732                      HEK-293CymR cell
#> 2973 BTO:0002974                     HEK-293-EBNA cell
#> 3031 BTO:0003032                         HEK-293F cell
#> 3032 BTO:0003033                         HEK-293H cell
#> 3181 BTO:0003182                             NHEK cell
#> 3262 BTO:0003263                        HEK-AD293 cell
#> 3346 BTO:0003347                HEK-293 Tet-On 3G cell
#> 3467 BTO:0003468                        HEK-293B2 cell
#> 3614 BTO:0003615                HEK-293 PEAKrapid cell
#> 4072 BTO:0004075                       D1-HEK-293 cell
#> 4789 BTO:0004792                           RHEK-1 cell
#> 5021 BTO:0005025                     HEK-293-TrkB cell
#> 5025 BTO:0005029                        HEK-293FT cell
#> 5544 BTO:0005548                        HEK-293ET cell
#> 5835 BTO:0005839                       HEK-293VnR cell
#> 6093 BTO:0006097                             HEKa cell
#> 6318 BTO:0006322             Jump-In-T-REx HEK293 cell
#> 6324 BTO:0006328                      HEK-293T/17 cell
#> 6418 BTO:0006422                   HEK-293-APPswe cell
#> 6449 BTO:0006453                         HEK-293S cell
#> 6509 BTO:0006513                      HEK-293-AT1 cell
searchSampleType(c("HEK", "SUM", "HeLa"), contain_all = FALSE)
#>               ID                                  Name
#> 567  BTO:0000567                             HeLa cell
#> 568  BTO:0000568                          HeLa-S3 cell
#> 2771 BTO:0002772                                 chela
#> 3179 BTO:0003180                          HeLa-80 cell
#> 3586 BTO:0003587                   HeLa-MAGI-CCR5 cell
#> 3587 BTO:0003588                        HeLa-MAGI cell
#> 3601 BTO:0003602             HeLa GFP-histone H2B cell
#> 5026 BTO:0005030                         HeLa-229 cell
#> 5034 BTO:0005038                HeLa-LTRHIV-1-Luc cell
#> 5995 BTO:0005999                           chelate leg
#> 6087 BTO:0006091                       HeLa-HL3T1 cell
#> 25   BTO:0000024                              abomasum
#> 273  BTO:0000272                     colon transversum
#> 349  BTO:0000348                                omasum
#> 362  BTO:0000361                    stratum granulosum
#> 436  BTO:0000435                      stratum spinosum
#> 615  BTO:0000615                       corpus callosum
#> 1712 BTO:0001713                                dorsum
#> 2017 BTO:0002018          corpus cavernosum clitoridis
#> 2018 BTO:0002019               corpus cavernosum penis
#> 3676 BTO:0003677                     chorion frondosum
#> 4133 BTO:0004136                         Kasumi-1 cell
#> 4422 BTO:0004425          stratum granulosum cerebelli
#> 4924 BTO:0004928                        SUM-159PT cell
#> 5383 BTO:0005387                        SUM-102PT cell
#> 5384 BTO:0005388                          SUM-149 cell
#> 5385 BTO:0005389                         SUM-44PE cell
#> 5386 BTO:0005390                         SUM-52PE cell
#> 5387 BTO:0005391                          SUM-225 cell
#> 5388 BTO:0005392                          SUM-229 cell
#> 5389 BTO:0005393                          SUM-190 cell
#> 5390 BTO:0005394                          SUM-185 cell
#> 5450 BTO:0005454                     corpus cavernosum
#> 5471 BTO:0005475                        SUM-149PT cell
#> 5472 BTO:0005476                        SUM-185PE cell
#> 5473 BTO:0005477                        SUM-190PT cell
#> 5474 BTO:0005478                       SUM-225CWN cell
#> 5475 BTO:0005479                        SUM-229PE cell
#> 5499 BTO:0005503       corpus cavernosum smooth muscle
#> 5928 BTO:0005932                              SUM cell
#> 5929 BTO:0005933                      SUM-1315MO2 cell
#> 6264 BTO:0006268 organum vasculosum laminae terminalis
#> 8    BTO:0000007                          HEK-293 cell
#> 2180 BTO:0002181                         HEK-293T cell
#> 2523 BTO:0002524                         HEK-293A cell
#> 2731 BTO:0002732                      HEK-293CymR cell
#> 2973 BTO:0002974                     HEK-293-EBNA cell
#> 3031 BTO:0003032                         HEK-293F cell
#> 3032 BTO:0003033                         HEK-293H cell
#> 3181 BTO:0003182                             NHEK cell
#> 3262 BTO:0003263                        HEK-AD293 cell
#> 3346 BTO:0003347                HEK-293 Tet-On 3G cell
#> 3467 BTO:0003468                        HEK-293B2 cell
#> 3614 BTO:0003615                HEK-293 PEAKrapid cell
#> 4072 BTO:0004075                       D1-HEK-293 cell
#> 4789 BTO:0004792                           RHEK-1 cell
#> 5021 BTO:0005025                     HEK-293-TrkB cell
#> 5025 BTO:0005029                        HEK-293FT cell
#> 5544 BTO:0005548                        HEK-293ET cell
#> 5835 BTO:0005839                       HEK-293VnR cell
#> 6093 BTO:0006097                             HEKa cell
#> 6318 BTO:0006322             Jump-In-T-REx HEK293 cell
#> 6324 BTO:0006328                      HEK-293T/17 cell
#> 6418 BTO:0006422                   HEK-293-APPswe cell
#> 6449 BTO:0006453                         HEK-293S cell
#> 6509 BTO:0006513                      HEK-293-AT1 cell
searchSampleType(c("HEK SUM", "HeLa"), contain_all = FALSE)
#>               ID                                  Name
#> 567  BTO:0000567                             HeLa cell
#> 568  BTO:0000568                          HeLa-S3 cell
#> 2771 BTO:0002772                                 chela
#> 3179 BTO:0003180                          HeLa-80 cell
#> 3586 BTO:0003587                   HeLa-MAGI-CCR5 cell
#> 3587 BTO:0003588                        HeLa-MAGI cell
#> 3601 BTO:0003602             HeLa GFP-histone H2B cell
#> 5026 BTO:0005030                         HeLa-229 cell
#> 5034 BTO:0005038                HeLa-LTRHIV-1-Luc cell
#> 5995 BTO:0005999                           chelate leg
#> 6087 BTO:0006091                       HeLa-HL3T1 cell
#> 25   BTO:0000024                              abomasum
#> 273  BTO:0000272                     colon transversum
#> 349  BTO:0000348                                omasum
#> 362  BTO:0000361                    stratum granulosum
#> 436  BTO:0000435                      stratum spinosum
#> 615  BTO:0000615                       corpus callosum
#> 1712 BTO:0001713                                dorsum
#> 2017 BTO:0002018          corpus cavernosum clitoridis
#> 2018 BTO:0002019               corpus cavernosum penis
#> 3676 BTO:0003677                     chorion frondosum
#> 4133 BTO:0004136                         Kasumi-1 cell
#> 4422 BTO:0004425          stratum granulosum cerebelli
#> 4924 BTO:0004928                        SUM-159PT cell
#> 5383 BTO:0005387                        SUM-102PT cell
#> 5384 BTO:0005388                          SUM-149 cell
#> 5385 BTO:0005389                         SUM-44PE cell
#> 5386 BTO:0005390                         SUM-52PE cell
#> 5387 BTO:0005391                          SUM-225 cell
#> 5388 BTO:0005392                          SUM-229 cell
#> 5389 BTO:0005393                          SUM-190 cell
#> 5390 BTO:0005394                          SUM-185 cell
#> 5450 BTO:0005454                     corpus cavernosum
#> 5471 BTO:0005475                        SUM-149PT cell
#> 5472 BTO:0005476                        SUM-185PE cell
#> 5473 BTO:0005477                        SUM-190PT cell
#> 5474 BTO:0005478                       SUM-225CWN cell
#> 5475 BTO:0005479                        SUM-229PE cell
#> 5499 BTO:0005503       corpus cavernosum smooth muscle
#> 5928 BTO:0005932                              SUM cell
#> 5929 BTO:0005933                      SUM-1315MO2 cell
#> 6264 BTO:0006268 organum vasculosum laminae terminalis
#> 8    BTO:0000007                          HEK-293 cell
#> 2180 BTO:0002181                         HEK-293T cell
#> 2523 BTO:0002524                         HEK-293A cell
#> 2731 BTO:0002732                      HEK-293CymR cell
#> 2973 BTO:0002974                     HEK-293-EBNA cell
#> 3031 BTO:0003032                         HEK-293F cell
#> 3032 BTO:0003033                         HEK-293H cell
#> 3181 BTO:0003182                             NHEK cell
#> 3262 BTO:0003263                        HEK-AD293 cell
#> 3346 BTO:0003347                HEK-293 Tet-On 3G cell
#> 3467 BTO:0003468                        HEK-293B2 cell
#> 3614 BTO:0003615                HEK-293 PEAKrapid cell
#> 4072 BTO:0004075                       D1-HEK-293 cell
#> 4789 BTO:0004792                           RHEK-1 cell
#> 5021 BTO:0005025                     HEK-293-TrkB cell
#> 5025 BTO:0005029                        HEK-293FT cell
#> 5544 BTO:0005548                        HEK-293ET cell
#> 5835 BTO:0005839                       HEK-293VnR cell
#> 6093 BTO:0006097                             HEKa cell
#> 6318 BTO:0006322             Jump-In-T-REx HEK293 cell
#> 6324 BTO:0006328                      HEK-293T/17 cell
#> 6418 BTO:0006422                   HEK-293-APPswe cell
#> 6449 BTO:0006453                         HEK-293S cell
#> 6509 BTO:0006513                      HEK-293-AT1 cell