BRENDASearch

search for ontology term in an obo file or a vector of all ontology term name updated 07/2020

BRENDASearch(x, obo = BRENDAobo, contain_all = TRUE)

Arguments

x
A string or character vector to search for. x will be divided into different search patterns by space. See examples below.

obo
The ontology file or a character vector of all possible term names. Optional. BRENDA OBO file will be loaded if not provided.

contain_all
if TRUE, will only return the results contain all search terms. if FALSE, will return results contain any of the given pattern.

Value

matrix including search result

Examples

BRENDASearch("MDA MB cell", contain_all = TRUE) 

#>             ID            Name                  
#> BTO:0000815 "BTO:0000815" "MDA-MB-231 cell"     
#> BTO:0001567 "BTO:0001567" "MDA-MB-435 cell"     
#> BTO:0001568 "BTO:0001568" "MDA-MB-436 cell"     
#> BTO:0001569 "BTO:0001569" "MDA-MB-361 cell"     
#> BTO:0001570 "BTO:0001570" "MDA-MB-468 cell"     
#> BTO:0001908 "BTO:0001908" "MDA-MB-453 cell"     
#> BTO:0003471 "BTO:0003471" "MDA-MB-175-VII cell" 
#> BTO:0003941 "BTO:0003941" "MDA-MB-435S cell"    
#> BTO:0004164 "BTO:0004164" "MDA-MB-435HGF cell"  
#> BTO:0004474 "BTO:0004474" "MDA-MB-231-BAG cell" 
#> BTO:0005172 "BTO:0005172" "MDA-MB-231-Luc2 cell"
#> BTO:0006039 "BTO:0006039" "MDA-MB-415 cell"     
#> BTO:0006218 "BTO:0006218" "MDA-MB-157 cell"     

# search for results that contain "MDA" "MB" and "cell"
BRENDASearch(c("MDA MB", "cell"), contain_all = TRUE) 

#>             ID            Name                  
#> BTO:0000815 "BTO:0000815" "MDA-MB-231 cell"     
#> BTO:0001567 "BTO:0001567" "MDA-MB-435 cell"     
#> BTO:0001568 "BTO:0001568" "MDA-MB-436 cell"     
#> BTO:0001569 "BTO:0001569" "MDA-MB-361 cell"     
#> BTO:0001570 "BTO:0001570" "MDA-MB-468 cell"     
#> BTO:0001908 "BTO:0001908" "MDA-MB-453 cell"     
#> BTO:0003471 "BTO:0003471" "MDA-MB-175-VII cell" 
#> BTO:0003941 "BTO:0003941" "MDA-MB-435S cell"    
#> BTO:0004164 "BTO:0004164" "MDA-MB-435HGF cell"  
#> BTO:0004474 "BTO:0004474" "MDA-MB-231-BAG cell" 
#> BTO:0005172 "BTO:0005172" "MDA-MB-231-Luc2 cell"
#> BTO:0006039 "BTO:0006039" "MDA-MB-415 cell"     
#> BTO:0006218 "BTO:0006218" "MDA-MB-157 cell"     

# same as above
BRENDASearch(c("MDA-MB", "SUM"), contain_all = FALSE) 

#>             ID            Name                  
#> BTO:0004928 "BTO:0004928" "SUM-159PT cell"      
#> BTO:0005387 "BTO:0005387" "SUM-102PT cell"      
#> BTO:0005389 "BTO:0005389" "SUM-44PE cell"       
#> BTO:0005390 "BTO:0005390" "SUM-52PE cell"       
#> BTO:0005475 "BTO:0005475" "SUM-149PT cell"      
#> BTO:0005476 "BTO:0005476" "SUM-185PE cell"      
#> BTO:0005477 "BTO:0005477" "SUM-190PT cell"      
#> BTO:0005478 "BTO:0005478" "SUM-225CWN cell"     
#> BTO:0005479 "BTO:0005479" "SUM-229PE cell"      
#> BTO:0005932 "BTO:0005932" "SUM cell"            
#> BTO:0005933 "BTO:0005933" "SUM-1315MO2 cell"    
#> BTO:0000815 "BTO:0000815" "MDA-MB-231 cell"     
#> BTO:0001567 "BTO:0001567" "MDA-MB-435 cell"     
#> BTO:0001568 "BTO:0001568" "MDA-MB-436 cell"     
#> BTO:0001569 "BTO:0001569" "MDA-MB-361 cell"     
#> BTO:0001570 "BTO:0001570" "MDA-MB-468 cell"     
#> BTO:0001908 "BTO:0001908" "MDA-MB-453 cell"     
#> BTO:0003471 "BTO:0003471" "MDA-MB-175-VII cell" 
#> BTO:0003941 "BTO:0003941" "MDA-MB-435S cell"    
#> BTO:0004164 "BTO:0004164" "MDA-MB-435HGF cell"  
#> BTO:0004474 "BTO:0004474" "MDA-MB-231-BAG cell" 
#> BTO:0005172 "BTO:0005172" "MDA-MB-231-Luc2 cell"
#> BTO:0006039 "BTO:0006039" "MDA-MB-415 cell"     
#> BTO:0006218 "BTO:0006218" "MDA-MB-157 cell"     

# search for results that contain either "MDA-MB" or "SUM"