GEOPlatformSearch

search for GEO platform name updated 02/2024

GEOPlatformSearch(
  x,
  organism = NULL,
  file = GEOplatform,
  contain_all = TRUE,
  accession_only = FALSE
)

Arguments

x

a string or character vector to search for. input character vector if searching for multiple terms.

organism

organism name to narrow down the search. partial input of an organism name is allowed. e.g. results for organism "Homo sapiens" will be pulled if input "homo".

file

GEO platform information dataframe. columns: "Accession", "Name", "Technology", "Organism".

contain_all

if TRUE, results contain all search terms will be returned. if FALSE, results contain any of the given term will be returned.

accession_only

if TRUE, only accession id of the search will be returned as a character. if FALSE, the complete information of the platforms will be returned as a dataframe.

Value

character or dataframe of search result

Examples

GEOPlatformSearch("illumina")
#>       Accession                                        Name
#> 3482   GPL30366        illumina novaseq 6000 (Mus musculus)
#> 3757   GPL30028        illumina NovaSeq 6000 (Mus musculus)
#> 3774   GPL30012             illumina NextSeq 500 (Zea mays)
#> 5361   GPL27991         illumina Nextseq 500 (Homo sapiens)
#> 7366   GPL24980           illumina Nextseq 500 (Bos taurus)
#> 7581   GPL24756  illumina HiSeq 2500 (Rhizophora mucronata)
#> 7848   GPL24432          illumina HiSeq 2500 (Homo sapiens)
#> 8826   GPL23177         illumina NextSeq 500 (Homo sapiens)
#> 8946   GPL23049 illumina Genome Analyzer IIx (Homo sapiens)
#> 11885  GPL19822          illumina HiSeq 2000 (Mus musculus)
#>                       Technology             Organism
#> 3482  high-throughput sequencing         Mus musculus
#> 3757  high-throughput sequencing         Mus musculus
#> 3774  high-throughput sequencing             Zea mays
#> 5361  high-throughput sequencing         Homo sapiens
#> 7366  high-throughput sequencing           Bos taurus
#> 7581  high-throughput sequencing Rhizophora mucronata
#> 7848  high-throughput sequencing         Homo sapiens
#> 8826  high-throughput sequencing         Homo sapiens
#> 8946  high-throughput sequencing         Homo sapiens
#> 11885 high-throughput sequencing         Mus musculus
GEOPlatformSearch("illumina", organism = "Homo sapiens")
#>      Accession                                        Name
#> 5361  GPL27991         illumina Nextseq 500 (Homo sapiens)
#> 7848  GPL24432          illumina HiSeq 2500 (Homo sapiens)
#> 8826  GPL23177         illumina NextSeq 500 (Homo sapiens)
#> 8946  GPL23049 illumina Genome Analyzer IIx (Homo sapiens)
#>                      Technology     Organism
#> 5361 high-throughput sequencing Homo sapiens
#> 7848 high-throughput sequencing Homo sapiens
#> 8826 high-throughput sequencing Homo sapiens
#> 8946 high-throughput sequencing Homo sapiens
GEOPlatformSearch(c("Affymetrix", "Transcriptome"), organism = "mus", contain_all = TRUE)
#>       Accession
#> 8543   GPL23535
#> 9943   GPL20266
#> 10013  GPL21877
#> 10497  GPL21341
#> 11015  GPL20775
#> 11503  GPL20258
#>                                                                                            Name
#> 8543  [MTA-1_0] Affymetrix Mouse Transcriptome Array 1.0 [CDF: Brainarray mta10_Mm_ENTREZG_v20]
#> 9943             [MTA-1_0] Affymetrix Mouse Transcriptome Array 1.0 [transcript (gene) version]
#> 10013                                        [MTA-1_0] Affymetrix Mouse Transcriptome Array 1.0
#> 10497          [MTA-1_0] Affymetrix Mouse Transcriptome Array 1.0 [probeset (exon) CDF version]
#> 11015        [MTA-1_0] Affymetrix Mouse Transcriptome Array 1.0 [transcript (gene) CSV version]
#> 11503        [MTA-1_0] Affymetrix Mouse Transcriptome Array 1.0 [transcript (gene) CDF version]
#>                    Technology     Organism
#> 8543  in situ oligonucleotide Mus musculus
#> 9943  in situ oligonucleotide Mus musculus
#> 10013 in situ oligonucleotide Mus musculus
#> 10497 in situ oligonucleotide Mus musculus
#> 11015 in situ oligonucleotide Mus musculus
#> 11503 in situ oligonucleotide Mus musculus