GEOPlatformSearch

search for GEO platform name updated 08/2020

GEOPlatformSearch(
  x,
  species = NULL,
  file = GEOplatform,
  contain_all = TRUE,
  accession_only = FALSE
)

Arguments

x
a string or character vector to search for. input character vector if searching for multiple terms.

species
species name to narrow down the search. partial input of an organism name is allowed. e.g. results for species "Homo sapiens" will be pulled if input "homo".

file
dataframe for platform infomation. contains "Title" to search for and "Taxnomy" for organism, and "Accession" for platform name. Optional. GEOplatform file will be loaded if not provided.

contain_all
if TRUE, results contain all search terms will be returned. if FALSE, results contain any of the given term will be returned.

accession_only
if TRUE, only accession id of the search will be returned as a character. if FALSE, the complete information of the platforms will be returned as a dataframe.

Value

character or dataframe of search result

Examples

GEOPlatformSearch("illumina")

#>      Accession                                       Title
#> 737   GPL27991         illumina Nextseq 500 (Homo sapiens)
#> 2763  GPL24980           illumina Nextseq 500 (Bos taurus)
#> 2979  GPL24756  illumina HiSeq 2500 (Rhizophora mucronata)
#> 3246  GPL24432          illumina HiSeq 2500 (Homo sapiens)
#> 4224  GPL23177         illumina NextSeq 500 (Homo sapiens)
#> 4344  GPL23049 illumina Genome Analyzer IIx (Homo sapiens)
#> 7283  GPL19822          illumina HiSeq 2000 (Mus musculus)
#>                      Technology             Taxonomy Data.Rows Samples.Count
#> 737  high-throughput sequencing         Homo sapiens         0             0
#> 2763 high-throughput sequencing           Bos taurus         0            36
#> 2979 high-throughput sequencing Rhizophora mucronata         0             1
#> 3246 high-throughput sequencing         Homo sapiens         0             8
#> 4224 high-throughput sequencing         Homo sapiens         0            13
#> 4344 high-throughput sequencing         Homo sapiens         0             2
#> 7283 high-throughput sequencing         Mus musculus         0             2
#>      Series.Count Contact Release.Date
#> 737             0     GEO Jan 13, 2020
#> 2763            2     GEO May 09, 2018
#> 2979            1     GEO Mar 21, 2018
#> 3246            2     GEO Dec 26, 2017
#> 4224            1     GEO Mar 14, 2017
#> 4344            1     GEO Feb 09, 2017
#> 7283            1     GEO Feb 24, 2015

GEOPlatformSearch("illumina", species = "homo sapiens")

#>      Accession                                       Title
#> 737   GPL27991         illumina Nextseq 500 (Homo sapiens)
#> 3246  GPL24432          illumina HiSeq 2500 (Homo sapiens)
#> 4224  GPL23177         illumina NextSeq 500 (Homo sapiens)
#> 4344  GPL23049 illumina Genome Analyzer IIx (Homo sapiens)
#>                      Technology     Taxonomy Data.Rows Samples.Count
#> 737  high-throughput sequencing Homo sapiens         0             0
#> 3246 high-throughput sequencing Homo sapiens         0             8
#> 4224 high-throughput sequencing Homo sapiens         0            13
#> 4344 high-throughput sequencing Homo sapiens         0             2
#>      Series.Count Contact Release.Date
#> 737             0     GEO Jan 13, 2020
#> 3246            2     GEO Dec 26, 2017
#> 4224            1     GEO Mar 14, 2017
#> 4344            1     GEO Feb 09, 2017

GEOPlatformSearch(c("Affymetrix", "Transcriptome"), species = "mus", contain_all = TRUE)

#>      Accession
#> 3941  GPL23535
#> 5341  GPL20266
#> 5411  GPL21877
#> 5895  GPL21341
#> 6413  GPL20775
#> 6901  GPL20258
#>                                                                                          Title
#> 3941 [MTA-1_0] Affymetrix Mouse Transcriptome Array 1.0 [CDF: Brainarray mta10_Mm_ENTREZG_v20]
#> 5341            [MTA-1_0] Affymetrix Mouse Transcriptome Array 1.0 [transcript (gene) version]
#> 5411                                        [MTA-1_0] Affymetrix Mouse Transcriptome Array 1.0
#> 5895          [MTA-1_0] Affymetrix Mouse Transcriptome Array 1.0 [probeset (exon) CDF version]
#> 6413        [MTA-1_0] Affymetrix Mouse Transcriptome Array 1.0 [transcript (gene) CSV version]
#> 6901        [MTA-1_0] Affymetrix Mouse Transcriptome Array 1.0 [transcript (gene) CDF version]
#>                   Technology     Taxonomy Data.Rows Samples.Count Series.Count
#> 3941 in situ oligonucleotide Mus musculus     25336            26            5
#> 5341 in situ oligonucleotide Mus musculus     80202           107            8
#> 5411 in situ oligonucleotide Mus musculus         0           289           17
#> 5895 in situ oligonucleotide Mus musculus    725678            79            8
#> 6413 in situ oligonucleotide Mus musculus     73070          1393          131
#> 6901 in situ oligonucleotide Mus musculus     72688           459           41
#>               Contact Release.Date
#> 3941        GEO admin Jun 01, 2017
#> 5341   Florent Dumont May 29, 2016
#> 5411 Affymetrix, Inc. May 18, 2016
#> 5895 Affymetrix, Inc. Jan 15, 2016
#> 6413 Affymetrix, Inc. Aug 05, 2015
#> 6901 Affymetrix, Inc. May 28, 2015